UCP/API: message priority field -> active message priority field, priority type uint8_t -> unsigned

+++ src/ucp/api/ucp.h
@@ -742,7 +742,7 @@ typedef enum {
                                                         of a single operation.
                                                         This flag and UCP_OP_ATTR_FLAG_FAST_CMPL
                                                         are mutually exclusive. */
-    UCP_OP_ATTR_FIELD_PRIORITY      = UCS_BIT(20)   /**< message priority field */
+    UCP_OP_ATTR_FIELD_PRIORITY      = UCS_BIT(20)   /**< active message priority field */
 } ucp_op_attr_t;
 
 
@@ -1269,7 +1269,7 @@ typedef struct ucp_worker_attr {
     /**
      * Maximum available message priority levels
      */
-    int                   max_message_priorities;
+    unsigned              max_message_priorities;
 } ucp_worker_attr_t;
 
 
@@ -1832,7 +1832,7 @@ typedef struct {
      * and depends on the capabilities of the selected transport,
      * ordering across messages with different priorities are also not guaranteed.
      */
-    uint8_t priority;
+    unsigned  priority;
 } ucp_request_param_t;
